backend/requirements.txt
# Flask and Core Dependencies
Flask>=2.3.3
Flask-SQLAlchemy>=3.0.5
Flask-Migrate>=4.0.5
Flask-CORS>=4.0.0
Werkzeug>=2.3.7

# Database
SQLAlchemy>=2.0.21
alembic>=1.12.0

# Excel Processing
pandas>=2.1.1
openpyxl>=3.1.2
xlrd>=2.0.1

# Data Processing and Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# HTTP and API
requests>=2.31.0
urllib3>=2.0.4

# Development and Testing
pytest>=7.4.2
pytest-flask>=1.2.0
pytest-cov>=4.1.0
black>=23.9.1
flake8>=6.1.0

# Production Server
gunicorn>=21.2.0

# Environment and Configuration
python-dotenv>=1.0.0
click>=8.1.7

# Logging and Monitoring
structlog>=23.1.0

# Security
cryptography>=41.0.4

# Optional: PostgreSQL support (for production)
# psycopg2-binary>=2.9.7

# Optional: Redis for caching (for production)
# redis>=5.0.1

# Optional: Celery for background tasks (for production)
# celery>=5.3.1 